+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+## Chanded
+- `insert` returns ID of the object inserted
+- `multiInsert` returns an `array` with ID of the objects inserted
+
 ## [1.3.0] - 2019-07-30
 ### Added
 - Order parameter for `get()` method